Aggiornamento

+++ Progetto/Master.c
@@ -1,6 +1,6 @@
 #include "Headers/master.h"
 #include "Headers/risorse.h"
-
+void terminazione(int idSemaforo,int idMemoriaCondivisa,int check, char* argv[],int argc);
 int flag = 0;
 int tempoScaduto = 0;
 int forkError = 0;
@@ -20,9 +20,15 @@ void handle_fork_error(int signal)
 {
     forkError = 1;
 }
+
 void main(int argc,char * argv[])
 {
-    printf("Master avviato\n");
+    /**
+     * Inizializzazione dei 3 hanlder per i seguenti segnali:
+     * 1) il primo serve per visualizzare ogni secondo le informazioni principali relative alla simulazione
+     * 2) il secondo viene inviato da un processo chiamato allarme e il segnale indica la terminazione per timeout
+     * 3) il terzo serve per segnale un errore durante una qualsiasi fork
+    */
     void handle_signal(int signal); /* the handler */
     void handle_exit(int signal);
     void handle_fork_error(int signal);
@@ -40,7 +46,10 @@ void main(int argc,char * argv[])
     s3.sa_handler = handle_fork_error;   /* set the handler */
     sigaction(SIGUSR2, &s3, NULL); /* CASE 1: set new handler */
 
-
+    /**
+     * Creazione dei processi seguenti
+     * Attivatore,Alimentatore
+    */
     pidAttivatore = fork();
     if (pidAttivatore == 0)
     {
@@ -61,18 +70,21 @@ void main(int argc,char * argv[])
 
     
     
-    printf("Attivatore avviato\nAlimentatore avviato\n");
-    struct sembuf my_op;
+    /**
+     * il metodo setSemaforo serve a creare ( oppure ad ottenere ) il vettore di semafori e inizializzarlo 
+    */
     int idSemaforo = setSemaforo();
-    my_op.sem_num = 0; /* only one semaphore in array of semaphores */
-    my_op.sem_flg = 0; /* no flag : default behavior */
-    my_op.sem_op = 1;  /* accessing the resource */
+   
 
     /**
      * Inizializzazione della memoria condivisa
      *
      */
     int idMemoriaCondivisa = setMemoriaCondivisa(N_ATOMI_INIT);
+
+    /**
+     * Creazione del processo inibitore
+    */
     if (argc > 1 && strcmp(argv[1], "inibitore") == 0)
     {
         
@@ -97,73 +109,111 @@ void main(int argc,char * argv[])
         
     }
     /**
-     * Creazione processi atomi iniziali
+     * Creazione primi N_ATOMI_INIT processi atomi iniziali
      */
     creaAtomi(N_ATOMI_INIT, N_ATOMO_MAX, idMemoriaCondivisa);
-    semop(idSemaforo, &my_op, 1);
+    struct sembuf my_op;
 
     /**
-     * CHANGELOG: il master non è più in sola lettura, ma deve aggiornare
-     * la memoria condivisa con l'energia prelevata, aggiunto controllo semaforo prioritario
-     */
-
-    my_op.sem_num = 1; /* only one semaphore in array of semaphores */
+     * Serve per dare il via a tutti i processi della simulazione
+    */
+    my_op.sem_num = 0; /* only one semaphore in array of semaphores */
     my_op.sem_flg = 0; /* no flag : default behavior */
     my_op.sem_op = 1;  /* accessing the resource */
     semop(idSemaforo, &my_op, 1);
 
-    my_op.sem_num = 2; /* only one semaphore in array of semaphores */
-    my_op.sem_flg = 0; /* no flag : default behavior */
-    my_op.sem_op = 1;  /* accessing the resource */
-    semop(idSemaforo, &my_op, 1);
 
-    /**
-     *Accesso alla memoria condivisa solo in lettura per la stampa, il master può accedervi in lettura
-     *solo per stampare le statistiche, quindi non è necessario controllare i semafori.
-     * CHANGELOG: non vale più, il master deve accedere alla memoria condivisa per aggiornare l'energia consumata
-     */
+    
     int check = 0;
+    /**
+     * Creazione del processo Alarm che si occupa soltanto di inviare il segnale al processo master che indica il timeout
+    */
     pidAllarme = fork();
     if (pidAllarme == 0)
     {
-        char *const array[2] = {"10", 0};
+        char *const array[1] = {0};
         execv("Alarm", array);
         perror("");
         exit(1);
     }
-    printf("Allarme avviato\n");
+
+
+
+    /**
+     * CORPO DEL CODICE
+    */
     while (check == 0 && tempoScaduto == 0 && forkError == 0)
     {
-
+        /**
+         * Periodicamente viene inviato un segnale per prelevare energia
+        */
         alarm(TIMER_PRELEVA);
+        /**
+         * Aspetta che i processi figli atomo terminino altrimenti diventano zombie
+        */
         wait(NULL);
 
+        /**
+         * Nel caso le 2 condizioni di uscita vengono soddisfatte, non ha senso proseguire e si procede a terminare la simulazione
+        */
         if (tempoScaduto == 0 && forkError ==  0)
         {
+            /**
+             * la variabile flag viene utilizzata per determinare se il segnale è stato ricevuto
+             * Nel caso ci fosse stato il segnale allora si procede a prelevare energia 
+             * altrimenti si aspetta con pause
+            */
             if (flag == 0)
             {
                 pause();
             }
+            /**
+             * Prima di entrare in sezione critica
+            */
             my_op.sem_num = 1; /* only one semaphore in array of semaphores */
             my_op.sem_flg = 0; /* no flag : default behavior */
             my_op.sem_op = -1; /* accessing the resource */
             semop(idSemaforo, &my_op, 1);
 
-            
+            //SEZIONE CRITICA: INIZIO
+
             prelevaEnergia(ENERGY_CONSUMPTION);
             stampa();
-
             struct memCond *datap = shmat(idMemoriaCondivisa, NULL, 0);
-            
+            /**
+             * Controlla se l'energia totale è inferiore a 0 oppure ha superato la soglia massima 
+            */
             check = checkEnergia();
 
+            //SEZIONE CRITICA: FINE
+
+            /*
+            *Rilascio del semaforo
+            */
             my_op.sem_op = 1; /* releasing the resource */
             semop(idSemaforo, &my_op, 1);
+            /**
+             *Il flag viene rimesso a 0 altrimenti riesegue la porzione di codice senza aspettare il segnale
+            */
             flag = 0;
         }
+        
     }
+
+terminazione(idSemaforo,idMemoriaCondivisa,check,argv,argc);
+
+
+    
+    
+}
+void terminazione(int idSemaforo,int idMemoriaCondivisa,int check, char* argv[],int argc){
+    struct sembuf my_op;
     kill(pidAlimentatore, SIGKILL);
-    kill(pidAttivatore, SIGHUP);
+    waitpid(pidAlimentatore,NULL,0);
+    printf("Il processo alimentatore ha terminato correttamente\n");
+    kill(pidAttivatore, SIGKILL);
+    waitpid(pidAttivatore,NULL,0);
+    printf("Il processo attivatore ha terminato correttamente\n");
     if (argc > 1 && strcmp(argv[1], "inibitore") == 0)
     {
         kill(pidInibitore, SIGKILL);
@@ -177,7 +227,12 @@ void main(int argc,char * argv[])
     int *array = shmat(idArrayCondiviso, NULL, 0);
     for (int i = 0; i < datap->nAtomi; i++)
     {
-        kill(array[i], SIGKILL);
+       if(array[i]!=-1){
+        kill(array[i],SIGKILL);
+        //wait(NULL);
+       }
+        
+        printf("Processo atomo con pid %d terminato correttamente\n",array[i]);
     }
 
     my_op.sem_num = 1; /* only one semaphore in array of semaphores */
@@ -226,5 +281,4 @@ void main(int argc,char * argv[])
     if(forkError == 1){
         printf("La simulazione è terminata per il seguente motivo: errore nelle fork dei processi\n");
     }
-    
 }
